requirements.txt
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
requests==2.32.3
tenacity==9.0.0
python-dotenv==1.0.1
